requirements.txt
openai==2.6.1
python-dotenv==1.2.1
streamlit==1.50.0